Windows port: op-family MSVC fixes (narrowing casts, builtins, force-inline)
- explicit casts where MSVC hard-errors on narrowing in braced initializers
  (GNU/Clang builds run these files with -Wno-narrowing): data_movement
  (concat, copy, gather, indexed_fill, moe remaps, roll, scatter, sort,
  transpose), eltwise (binary_ng, ternary, unary), matmul factories,
  reduction (accumulation, topk), core work_split
- __builtin_unreachable() call sites routed through ttsl::unreachable()
- __builtin_clz replaced with std::bit_floor in topk_utils
- bare glibc 'uint' spelled uint32_t in unary_op_utils
- slice_device_operation uses TT_FORCE_INLINE instead of raw
  __attribute__((always_inline))

(cherry picked from commit 605eba2eeed5ecad58cbdd581f2437522d2d79c3)

+++ ttnn/cpp/ttnn/operations/core/work_split/work_split_tilize.hpp
@@ -107,7 +107,8 @@ inline NcoresWH compute_ncores_wh(size_t grid_area, uint32_t nblocks, uint32_t w
     const uint32_t total_blocks_height = tt::div_up(height_tiles, single_block_size);
     const uint32_t total_blocks = total_blocks_width * total_blocks_height;
     const uint32_t ncores = (nblocks_per_core == 0) ? nblocks : total_blocks;
-    return NcoresWH{ncores, nblocks_per_core, total_blocks_width, total_blocks_height, single_block_size};
+    return NcoresWH{
+        ncores, nblocks_per_core, total_blocks_width, total_blocks_height, static_cast<uint32_t>(single_block_size)};
 }
 
 // Non-throwing: returns nullopt when no sub-block assignment satisfies the L1-derived
